Refactor sqlmodel module to replace old db module
